Restore the Legacy tree where the merge silently took dev's side.
A plain merge only asks for a decision when both sides touched a file. Where
only dev moved, git took dev's version without comment. Across a divergence
this wide that is the wrong default: it pulled in dev's
Geometry/Primitives/Interfaces.cs, whose interface shape no longer matches
Legacy's GridVector2, GridVector3 and friends, and Geometry stopped compiling
with twenty errors. The same thing happened quietly to 134 other files the
plan assigns to Legacy, across the client, AnnotationVizLib, MorphologyMesh
and the EF Core model.

Every path outside dev's areas is now byte-identical to recovery-merge, so
later phases port dev's work on purpose instead of inheriting it by accident.

Three shared projects also go back to Legacy for now. dev's
gRPCAnnotationServiceTypes, WebAnnotationModel and GrpcAnnotationService all
compile against dev's AnnotationInterfaces, which is a different shape from
Legacy's and is referenced by fourteen projects including the entire client.
Swapping it is Phase 4's job, not something to do by merge default, so these
build on Legacy's interfaces until then.

WebAnnotationModel.gRPC, its tests, and WebAnnotationModel.WCF stay in the
tree but come back out of Everything.sln, because they cannot compile until
that same interface reconciliation happens. Nothing is lost: every version
mentioned here is reachable from the dev parent of the merge and from
origin/dev.

Verified against the pre-merge tree: thirteen representative projects build
with the same error counts or better, and SegmentationServiceTypes.gRPC now
builds clean where it previously failed, because its proto submodule is
finally declared and checked out.

+++ Servers/AnnotationService/AnnotationService/Service/Types/StructureTypeExtensions.cs
@@ -47,7 +47,7 @@ namespace AnnotationService.Types
             st.StructureTags = type.StructureTags is null ? [] : type.StructureTags.Split(';');
 
             st.Abstract = type.Abstract;
-            st.Color = (uint)type.Color;
+            st.Color = type.Color;
             st.Code = type.Code;
             st.HotKey = type.HotKey.Length > 0 ? type.HotKey[0] : '\0';
             st.PermittedLinks = PopulateLinks(type);
@@ -79,7 +79,7 @@ namespace AnnotationService.Types
 
             type.StructureTags = structuretags;
             type.Abstract = st.Abstract;
-            type.Color = (int)st.Color;
+            type.Color = st.Color;
             type.Code = st.Code;
             type.HotKey = st.HotKey.ToString();
             type.Username = Annotation.ServiceModelUtil.GetUserForCall();
